# Conflicts:
#	Cargo.lock
#	ic-os/defs.bzl
#	rs/consensus/dkg/src/lib.rs
#	rs/consensus/dkg/src/payload_validator.rs
#	rs/consensus/mocks/src/lib.rs
#	rs/consensus/src/consensus.rs
#	rs/consensus/src/consensus/block_maker.rs
#	rs/consensus/src/consensus/validator.rs
#	rs/consensus/tests/framework/runner.rs
#	rs/consensus/tests/integration.rs
#	rs/consensus/tests/payload.rs
#	rs/consensus/utils/src/lib.rs
#	rs/determinism_test/src/setup.rs
#	rs/http_endpoints/public/src/status.rs
#	rs/https_outcalls/consensus/src/pool_manager.rs
#	rs/replay/src/validator.rs
#	rs/replica/bin/replica/main.rs
#	rs/replica/setup_ic_network/src/lib.rs
#	rs/replica/src/setup.rs
#	rs/replica/src/setup_ic_stack.rs
#	rs/replica_tests/src/lib.rs
#	rs/state_machine_tests/src/lib.rs
#	rs/types/types/src/replica_config.rs

+++ rs/nervous_system/common/src/lib.rs
@@ -73,7 +73,7 @@ pub const START_OF_2022_TIMESTAMP_SECONDS: u64 = 1641016800;
 pub const ONE_TRILLION: u64 = 1_000_000_000_000;
 
 /// The number of cycles required to create an SNS, charged by the SNS-W canister.
-pub const SNS_CREATION_FEE: u64 = 180 * ONE_TRILLION;
+pub const SNS_CREATION_FEE: u128 = 180 * ONE_TRILLION as u128;
 
 // The number of nanoseconds per second.
 pub const NANO_SECONDS_PER_SECOND: u64 = 1_000_000_000;
